refactor(edit): rename VehicleStartLocationOverlay to AddressOverlay for future reuse with delivery cards
